# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory API Documentation",
        "description": "API documentation for the Inventory project",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Inventory API Server"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lister toutes les catégories",
                "description": "Récupère la liste des catégories associées à l'entreprise de l'utilisateur connecté.",
                "operationId": "0a3810ef08509327a6db442d3e4c4b73",
                "responses": {
                    "200": {
                        "description": "Succès - Liste récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categorie"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "f44066b315f1a6670d352e128ba39465",
                "requestBody": {
                    "description": "Données de la nouvelle catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_cat",
                                    "type"
                                ],
                                "properties": {
                                    "name_cat": {
                                        "description": "Nom de la catégorie",
                                        "type": "string",
                                        "example": "Ordinateurs Portables"
                                    },
                                    "type": {
                                        "description": "Type de la catégorie",
                                        "type": "string",
                                        "example": "Matériel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès - Catégorie créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Afficher les détails d'une catégorie",
                "operationId": "829dd0fbc1687efece34524110808262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Détails de la catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Catégorie introuvable ou accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mettre à jour une catégorie",
                "operationId": "68cf6ef85d9b79cea99f6063deca74e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Champs à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_cat": {
                                        "type": "string",
                                        "example": "Imprimantes Laser"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Bureautique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès - Catégorie mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "c637fcbceb0196c4d850f96245587511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Succès - Catégorie supprimée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categorie": {
                "title": "Categorie",
                "description": "Catégorie de produits",
                "properties": {
                    "id_categorie": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_entreprise": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_users": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name_cat": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "entreprise": {
                        "$ref": "#/components/schemas/Entreprise"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Entreprise": {
                "title": "Entreprise",
                "description": "Informations sur une entreprise",
                "properties": {
                    "id_entreprise": {
                        "description": "Identifiant unique de l'entreprise",
                        "type": "string",
                        "format": "uuid"
                    },
                    "company_name": {
                        "description": "Nom de l'entreprise",
                        "type": "string",
                        "maxLength": 255
                    },
                    "tax_number": {
                        "description": "Numéro fiscal",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "description": "Adresse de l'entreprise",
                        "type": "string",
                        "nullable": true
                    },
                    "company_email": {
                        "description": "Email de l'entreprise",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "logo": {
                        "description": "Logo (URL ou chemin)",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_name": {
                        "description": "Nom du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_email": {
                        "description": "Email du responsable",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "manager_phone": {
                        "description": "Téléphone du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_nui": {
                        "description": "NUI du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Profil": {
                "title": "Profil",
                "properties": {
                    "id_profil": {
                        "description": "Profile's unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom": {
                        "description": "Profile's name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Profile's description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Utilisateur du système",
                "properties": {
                    "id_users": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_entreprise": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "agence": {
                        "type": "string",
                        "nullable": true
                    },
                    "poste": {
                        "type": "string",
                        "nullable": true
                    },
                    "link_img": {
                        "type": "string",
                        "nullable": true
                    },
                    "matricule": {
                        "type": "string",
                        "nullable": true
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "entreprise": {
                        "$ref": "#/components/schemas/Entreprise"
                    },
                    "profil": {
                        "$ref": "#/components/schemas/Profil"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}